+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Reserved qty: Add data checks in service and in view.
 - Sale Order Line: When being edited, editable when the amount is 0.
 - ManufOrder: Prevent the user from removing realized stock move lines.
+- ManufOrder: does not create empty stock moves.
 
 ## Bug Fixes
 - Configurator : Fix demo data
@@ -18,6 +19,7 @@
 - BankPayment : Change in slice of nextOrderId in EbicsUser.
 - MANUF. ORDER : display message when no default email account found to send email
 - STOCK MOVE : address printing issue
+- Ticket : remove pop-up onNew.
 - Product : set 'stockManaged' false when product is of type 'service'.
 
 ## [5.1.4] - 2019-03-11